moved fallback component
