+++ frontend/src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {RouterModule} from '@angular/router';
-import {FallbackComponent} from './fallback/fallback.component';
+import {FallbackComponent} from './main/fallback/fallback.component';
 import {AuthGuard} from './core/services/auth/auth-guard.service';
 
 @NgModule({
