+++ frontend/src/app/main/fallback/fallback.component.scss
